added thread struct

+++ srcs/complex.c
@@ -6,13 +6,13 @@
 /*   By: aledru <aledru@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/01/17 14:22:17 by aledru            #+#    #+#             */
-/*   Updated: 2018/01/25 14:29:51 by aledru           ###   ########.fr       */
+/*   Updated: 2018/01/26 13:35:43 by aledru           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "fractol.h"
 
-t_complex	*create_complex(float r, float i)
+t_complex	*create_complex(double r, double i)
 {
 	t_complex	*complex;
 
@@ -23,7 +23,7 @@ t_complex	*create_complex(float r, float i)
 	return (complex);
 }
 
-void		set_complex(t_complex *complex, float r, float i)
+void		set_complex(t_complex *complex, double r, double i)
 {
 	complex->r = r;
 	complex->i = i;
